+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Add `salmon.util/pages-seq` for handling paginated responses.
 - Add `salmon.cleanup/deregister-all-amis!` for cleaning up AMIs in test accounts.
+- Add `salmon.cleanup/delete-orphaned-snapshots!` for cleaning up snapshots in test accounts.
 
 ## v0.19.0 (2024-09-16)